This commit adds two important shims:
FIrst it adds compatibility functions (as appropriate) so that we can
support PostgreSQL 9.6 and 10+ at the same time with the same
interface.

Secondly it adds an indirection layer around the wal telemetry routines
so that they can be safely run on the replicas as well as the master
although the implications of doing so are different (because we look
at replay rather than write locations).