Refactoring, new view, bump up to 1.2
Add a view to show all client backends waiting on a lock for more than 5
minutes. The goal is to have a graph and a check reporting all long
client backend locks

Slight refactoring, use is_replica instead of pg_is_in_recovery()
consistently, use a saner way of checking Posgres versions, put
pgtelemetry-head in line with the latest version, use
client_stat_activity instead of pg_stat_activity in connections_by_ip
view, as the caller is clearly not interested in autovacuums and
friends.

Bump up to 1.2, update tests and migration files, put a version of the
expected file  for testing locally when pg_stat_statements is not in the
shared_preload_libraries.